Compression & finalization
This is the "final" patch for the redCubeAuton.h that I developed. This
is known as "Ben's Trash", but I feel that this program is ready for
implementation.

It can be easly adapted to suit any auton that you have written since I
started working.
