+++ State/includes.h
@@ -9,3 +9,7 @@
 #include "lnBatt.h"
 
 #include "Autons/autons.h"
+
+#include "redCubeAuton.h"
+
+#include "Autons/Auton Functions.h"
